api-compras/backend/requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
email-validator>=2.0    
pydantic>=2.8,<3
alembic>=1.13.0
werkzeug>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9 
bcrypt==4.0.1
PyJWT==2.8.0
httpx>=0.24.0
psycopg2-binary>=2.9.9
